Prod deploy

+++ api/src/cron/fetchRates.ts
@@ -6,17 +6,24 @@ import { __dirname } from "../config";
 export const cron_fetchRates = async () => {
   const tempDir = path.join(__dirname, `/../temp`);
 
+  // Don't run in dev
+  if (process.env.NODE_ENV === "development") {
+    return;
+  }
+
   // Remove old rates
   for (const file of await fs.readdir(tempDir)) {
     if (file === ".gitkeep") continue; // Keep on git
     if (file === "media") continue; // For media files
 
-    await fs.unlink(path.join(tempDir, file));
+    if (file === "exchange_rates.json") {
+      await fs.unlink(path.join(tempDir, file));
+    }
   }
 
   // Make sure that API key exists
   if (!process.env.OPEN_EXCHANGE_RATES_API_KEY) {
-    // console.log("Missing OXR API key");
+    console.log("cron_fetchRates failed: Missing OXR API key");
     return;
   }
